requirements.txt
Cerberus==1.1
Flask==0.11.1
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.1
gevent==1.1.2
greenlet==0.4.11
Jinja2==2.8
PyMySQL==0.7.9
redis==2.10.5
requests==2.12.3
grequests==0.3.0
selenium==3.11.0
gunicorn==19.7.1